Factored out function loading/ compilation
