+++ src/wasm/CMakeLists.txt
@@ -5,6 +5,7 @@ include_directories(
 
 add_library(wasm STATIC
         CallChain.cpp
+        FunctionLoader.cpp
         WasmModule.cpp
         syscalls.cpp
         ${FAASM_INCLUDE_DIR}/wasm/CallChain.h
