+++ src/wasm/WasmModule.cpp
@@ -4,14 +4,9 @@
 #include <util/func.h>
 #include <util/memory.h>
 #include <prof/prof.h>
+#include <wasm/FunctionLoader.h>
 
 #include <syscall.h>
-
-#include <WAVM/WASM/WASM.h>
-#include <WAVM/Inline/CLI.h>
-#include <WAVM/IR/Types.h>
-#include <WAVM/Runtime/RuntimeData.h>
-
 
 using namespace WAVM;
 
@@ -43,7 +38,6 @@ namespace wasm {
 
     WasmModule::~WasmModule() {
         // delete[] cleanMemory;
-
         defaultMemory = nullptr;
         moduleInstance = nullptr;
         functionInstance = nullptr;
@@ -95,9 +89,14 @@ namespace wasm {
             throw std::runtime_error("Cannot bind a module twice");
         }
 
-        // Parse the wasm file to work out imports, function signatures etc.
+        // Load the function data
         const util::TimePoint &wasmParseTs = prof::startTimer();
-        this->parseWasm(msg);
+        FunctionLoader f;
+        f.loadFunctionBytes(msg, module);
+
+        // Set up minimum memory size
+        module.memories.defs[0].type.size.min = (U64) INITIAL_MEMORY_PAGES;
+
         prof::logEndTimer("wasm-parse", wasmParseTs);
 
         // Linking
@@ -234,45 +233,6 @@ namespace wasm {
             exitCode = e.exitCode;
         }
         return exitCode;
-    }
-
-    std::vector<uint8_t> WasmModule::compile(message::Message &msg) {
-        // Parse the wasm file to work out imports, function signatures etc.
-        WasmModule tempModule;
-        tempModule.parseWasm(msg);
-
-        // Compile the module to object code
-        Runtime::ModuleRef module = Runtime::compileModule(tempModule.module);
-        return Runtime::getObjectCode(module);
-    }
-
-    void WasmModule::compileToObjectFile(message::Message &msg) {
-        std::vector<uint8_t> objBytes = wasm::WasmModule::compile(msg);
-        std::string objFilePath = util::getFunctionObjectFile(msg);
-        util::writeBytesToFile(objFilePath, objBytes);
-    }
-
-    /**
-     * Parse the WASM file to work out functions, exports, imports etc.
-     */
-    void WasmModule::parseWasm(const message::Message &msg) {
-        const std::shared_ptr<spdlog::logger> &logger = util::getLogger();
-
-        std::vector<U8> fileBytes;
-        std::string filePath = util::getFunctionFile(msg);
-
-        if (!loadFile(filePath.c_str(), fileBytes)) {
-            logger->error("Could not read data from {}", filePath);
-            throw std::runtime_error("Could not read binary data from file");
-        }
-
-        WASM::loadBinaryModule(fileBytes.data(), fileBytes.size(), this->module);
-
-        // Set up minimum memory size
-        this->module.memories.defs[0].type.size.min = (U64) INITIAL_MEMORY_PAGES;
-
-        // Add the shared memory definition
-        // this->module.memories.imports.push_back({{true, {0, 1000}}, msg.user(), "shared_state"});
     }
 
     U32 WasmModule::mmap(U32 length) {
